Fix metascraper-date's incorrect parsing of dates in certain conditions
Here is an example page which currently has an incorrectly parsed date:

https://www.health.govt.nz/about-ministry/legislation-and-regulation/regulatory-impact-statements/quality-improvement-agency-health-quality-and-safety-commission-functions-powers-and-funding

The HTML which is extracted to find the date string is

```html
<p class="page_updated">Page last updated: <span class="date">07 April 2011</span></p>
```

Currently the extracted date is `201107-04-07` because the selector
matches multiple elements which overlap resulting in the extracted text
being `Page last updated: 07 April 201107 April 2011`.

This pull request fixes this bug by mapping over each matched element
and attempting to parse it as a date, returning the first successful
match from that selector. With the fix, the date on this page is
correctly parsed as `2017-04-07`.