perf(logo-favicon): memoize urls resolution (#512)
Using a function memoizer. This is a necessary thing in this case because http agent only cache succesful/redirect responses.

In this very particular cases, the most common scenario is favicon is missing (404).

Since keyv doesn't store `undefined`, in that case we convert it to `null` and then back to `undefined`.